+++ CHANGES.rst
@@ -5,6 +5,8 @@ Changes For IRIS Service
 unreleased
 ==========
 
+ - access permissions for petition events:
+   all events need admin permissions except `publish` and `support`
  - added general petition extender to remove private data from the owner
    relation.